rag_api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pymongo>=4.6.0
requests>=2.31.0
numpy>=1.24.0
faiss-cpu>=1.7.4
datasets>=2.14.0

